pyproject.toml
[tool.poetry]
name = "socket_stream"
version = "0.1.0"
description = ""
authors = ["Chumakova Anastasia"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
pyzmq = "^23.2.0"
opencv-python = "^4.6.0"
pydantic = {extras = ["email", "dotenv"], version = "^1.9.1"}
numpy = "^1.23.1"
Flask = "^2.2.1"
loguru = "^0.6.0"
art = "^5.7"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
my-stream = "src.main:main"
